Moved Examples to dedicated package + testing.
